Setup of eslint, jest and Github Action
Copy webapp folder into exercise 2 directory and install eslint with prettier and eslint-plugin-jest, jest and vue-test-utils
and add a file to devine Github Actions. Jest is configured to also find tests in the src directory and eslint recognises them
there as such. The configurations for eslint and jest are done in the package.json.